chore: prefer generic validation functions, clean up
